Changes to types
Various small changes, but most notably to adjust the syntactic type
system so that it is consistent with Montague’s PTQ. That is, a
functional type A/B is understood as a function from B to A (not
vice-versa). We seemed to be doing both versions in different bits of
the paper!